wtp/webrtc/RTCPeerConnection-getStats.https.html

<!doctype html>
<meta charset="utf-8">
<title>RTCPeerConnection.prototype.getStats</title>
<script src="/resources/webrtc-ie-shim.js"></script><script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script src="dictionary-helper.js"></script>
<script src="RTCStats-helper.js"></script>
<script>
  'use strict';
  
  // Test is based on the following editor draft:
  // webrtc-pc 20171130
  // webrtc-stats 20171122
  
  // The following helper function is called from RTCPeerConnection-helper.js
  //   getTrackFromUserMedia
  
  // The following helper function is called from RTCStats-helper.js
  //   validateStatsReport
  //   assert_stats_report_has_stats
  
  // The following helper function is called from RTCPeerConnection-helper.js
  //   exchangeIceCandidates
  //   doSignalingHandshake
  
  /*
    8.2.  getStats
      1.  Let selectorArg be the method's first argument.
      2.  Let connection be the RTCPeerConnection object on which the method was invoked.
      3.  If selectorArg is null, let selector be null.
      4.  If selectorArg is a MediaStreamTrack let selector be an RTCRtpSender
          or RTCRtpReceiver on connection which track member matches selectorArg.
          If no such sender or receiver exists, or if more than one sender or
          receiver fit this criteria, return a promise rejected with a newly
          created InvalidAccessError.
      5.  Let p be a new promise.
      6.  Run the following steps in parallel:
        1.  Gather the stats indicated by selector according to the stats selection algorithm.
        2.  Resolve p with the resulting RTCStatsReport object, containing the gathered stats.
   */
  
  var _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"]) _i["return"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError("Invalid attempt to destructure non-iterable instance"); } }; }();
  
  promise_test(function () {
    var pc = new RTCPeerConnection();
    return pc.getStats();
  }, 'getStats() with no argument should succeed');
  
  promise_test(function () {
    var pc = new RTCPeerConnection();
    return pc.getStats(null);
  }, 'getStats(null) should succeed');
  
  /*
    8.2.  getStats
      4.  If selectorArg is a MediaStreamTrack let selector be an RTCRtpSender
          or RTCRtpReceiver on connection which track member matches selectorArg.
          If no such sender or receiver exists, or if more than one sender or
          receiver fit this criteria, return a promise rejected with a newly
          created InvalidAccessError.
   */
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    return getTrackFromUserMedia('audio').then(function (_ref) {
      var _ref2 = _slicedToArray(_ref, 2),
          track = _ref2[0],
          mediaStream = _ref2[1];
  
      return promise_rejects(t, 'InvalidAccessError', pc.getStats(track));
    });
  }, 'getStats() with track not added to connection should reject with InvalidAccessError');
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    return getTrackFromUserMedia('audio').then(function (_ref3) {
      var _ref4 = _slicedToArray(_ref3, 2),
          track = _ref4[0],
          mediaStream = _ref4[1];
  
      pc.addTrack(track, mediaStream);
      return pc.getStats(track);
    });
  }, 'getStats() with track added via addTrack should succeed');
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    var track = generateMediaStreamTrack();
    pc.addTransceiver(track);
  
    return pc.getStats(track);
  }, 'getStats() with track added via addTransceiver should succeed');
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    return getTrackFromUserMedia('audio').then(function (_ref5) {
      var _ref6 = _slicedToArray(_ref5, 2),
          track = _ref6[0],
          mediaStream = _ref6[1];
  
      // addTransceiver allows adding same track multiple times
      var transceiver1 = pc.addTransceiver(track);
      var transceiver2 = pc.addTransceiver(track);
  
      assert_not_equals(transceiver1, transceiver2);
      assert_not_equals(transceiver1.sender, transceiver2.sender);
      assert_equals(transceiver1.sender.track, transceiver2.sender.track);
  
      return promise_rejects(t, 'InvalidAccessError', pc.getStats(track));
    });
  }, 'getStats() with track associated with more than one sender should reject with InvalidAccessError');
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    var transceiver1 = pc.addTransceiver('audio');
  
    // Create another transceiver that resends what
    // is being received, kind of like echo
    var transceiver2 = pc.addTransceiver(transceiver1.receiver.track);
    assert_equals(transceiver1.receiver.track, transceiver2.sender.track);
  
    return promise_rejects(t, 'InvalidAccessError', pc.getStats(transceiver1.receiver.track));
  }, 'getStats() with track associated with both sender and receiver should reject with InvalidAccessError');
  
  /*
    8.5.  The stats selection algorithm
      2.  If selector is null, gather stats for the whole connection, add them to result,
          return result, and abort these steps.
   */
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    return pc.getStats().then(function (statsReport) {
      validateStatsReport(statsReport);
      assert_stats_report_has_stats(statsReport, ['peer-connection']);
    });
  }, 'getStats() with no argument should return stats report containing peer-connection stats on an empty PC');
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    return getTrackFromUserMedia('audio').then(function (_ref7) {
      var _ref8 = _slicedToArray(_ref7, 2),
          track = _ref8[0],
          mediaStream = _ref8[1];
  
      pc.addTrack(track, mediaStream);
      return pc.getStats();
    }).then(function (statsReport) {
      validateStatsReport(statsReport);
      assert_stats_report_has_stats(statsReport, ['peer-connection']);
      assert_stats_report_has_stats(statsReport, ['outbound-rtp']);
    });
  }, 'getStats() with no argument should return stats report containing peer-connection stats and outbound-track-stats');
  
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    return getTrackFromUserMedia('audio').then(function (_ref9) {
      var _ref10 = _slicedToArray(_ref9, 2),
          track = _ref10[0],
          mediaStream = _ref10[1];
  
      pc.addTrack(track);
      return pc.getStats();
    }).then(function (statsReport) {
      validateStatsReport(statsReport);
      assert_stats_report_has_stats(statsReport, ['peer-connection']);
      assert_stats_report_has_stats(statsReport, ['outbound-rtp']);
    });
  }, 'getStats() with no argument should return stats for no-stream tracks');
  
  /*
    8.5.  The stats selection algorithm
      3.  If selector is an RTCRtpSender, gather stats for and add the following objects
          to result:
        - All RTCOutboundRTPStreamStats objects corresponding to selector.
        - All stats objects referenced directly or indirectly by the RTCOutboundRTPStreamStats
          objects added.
   */
  promise_test(function () {
    var pc = new RTCPeerConnection();
    return getTrackFromUserMedia('audio').then(function (_ref11) {
      var _ref12 = _slicedToArray(_ref11, 2),
          track = _ref12[0],
          mediaStream = _ref12[1];
  
      pc.addTrack(track, mediaStream);
  
      return pc.getStats(track).then(function (statsReport) {
        validateStatsReport(statsReport);
        assert_stats_report_has_stats(statsReport, ['outbound-rtp']);
      });
    });
  }, 'getStats() on track associated with RtpSender should return stats report containing outbound-rtp stats');
  
  /*
    8.5.  The stats selection algorithm
      4.  If selector is an RTCRtpReceiver, gather stats for and add the following objects
          to result:
        - All RTCInboundRTPStreamStats objects corresponding to selector.
        - All stats objects referenced directly or indirectly by the RTCInboundRTPStreamStats
          added.
   */
  promise_test(function () {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
  
    return pc.getStats(transceiver.receiver.track).then(function (statsReport) {
      validateStatsReport(statsReport);
      assert_stats_report_has_stats(statsReport, ['inbound-rtp']);
    });
  }, 'getStats() on track associated with RtpReceiver should return stats report containing inbound-rtp stats');
  
  /*
    8.6   Mandatory To Implement Stats
      An implementation MUST support generating statistics of the following types
      when the corresponding objects exist on a PeerConnection, with the attributes
      that are listed when they are valid for that object.
   */
  
  var mandatoryStats = ["codec", "inbound-rtp", "outbound-rtp", "remote-inbound-rtp", "remote-outbound-rtp", "peer-connection", "data-channel", "stream", "track", "transport", "candidate-pair", "local-candidate", "remote-candidate", "certificate"];
  
  async_test(function (t) {
    var pc1 = new RTCPeerConnection();
    t.add_cleanup(function () {
      return pc1.close();
    });
    var pc2 = new RTCPeerConnection();
    t.add_cleanup(function () {
      return pc2.close();
    });
  
    var dataChannel = pc1.createDataChannel('test-channel');
  
    return navigator.mediaDevices.getUserMedia({
      audio: true,
      video: true
    }).then(t.step_func(function (mediaStream) {
      var tracks = mediaStream.getTracks();
      assert_equals(tracks.length, 2, 'Expect media stream to have one audio and one video track');
  
      var audioTrack = void 0;
      var videoTrack = void 0;
  
      var _iteratorNormalCompletion = true;
      var _didIteratorError = false;
      var _iteratorError = undefined;
  
      try {
        var _loop = function _loop() {
          var track = _step.value;
  
          t.add_cleanup(function () {
            return track.stop();
          });
  
          pc1.addTrack(track, mediaStream);
  
          if (track.kind === 'audio') {
            audioTrack = track;
          } else if (track.kind === 'video') {
            videoTrack = track;
          }
        };
  
        for (var _iterator = tracks[Symbol.iterator](), _step; !(_iteratorNormalCompletion = (_step = _iterator.next()).done); _iteratorNormalCompletion = true) {
          _loop();
        }
      } catch (err) {
        _didIteratorError = true;
        _iteratorError = err;
      } finally {
        try {
          if (!_iteratorNormalCompletion && _iterator.return) {
            _iterator.return();
          }
        } finally {
          if (_didIteratorError) {
            throw _iteratorError;
          }
        }
      }
  
      if (!audioTrack || !videoTrack) {
        assert_unreached('Expect mediaStream to have both audio and video streams');
      }
  
      var testStatsReport = function testStatsReport(pc, statsReport) {
        validateStatsReport(statsReport);
        assert_stats_report_has_stats(statsReport, mandatoryStats);
  
        var dataChannelStats = findStatsFromReport(statsReport, function (stats) {
          return stats.type === 'data-channel' && stats.dataChannelIdentifier === dataChannel.id;
        }, 'Expect data channel stats to be found');
  
        assert_equals(dataChannelStats.label, 'test-channel');
  
        var audioTrackStats = findStatsFromReport(statsReport, function (stats) {
          return stats.type === 'track' && stats.trackIdentifier === audioTrack.id;
        }, 'Expect audio track stats to be found');
  
        assert_equals(audioTrackStats.kind, 'audio');
  
        var videoTrackStats = findStatsFromReport(statsReport, function (stats) {
          return stats.type === 'track' && stats.trackIdentifier === videoTrack.id;
        }, 'Expect video track stats to be found');
  
        assert_equals(videoTrackStats.kind, 'video');
  
        var mediaStreamStats = findStatsFromReport(statsReport, function (stats) {
          return stats.type === 'stream' && stats.streamIdentifier === mediaStream.id;
        }, 'Expect media stream stats to be found');
  
        assert_true(mediaStreamStats.trackIds.include(audioTrackStats.id));
        assert_true(mediaStreamStats.trackIds.include(videoTrackStats.id));
      };
  
      var onConnected = t.step_func(function () {
        // Wait a while for the peer connections to collect stats
        t.step_timeout(function () {
          Promise.all([pc1.getStats().then(function (statsReport) {
            return testStatsReport(pc1, statsReport);
          }), pc2.getStats().then(function (statsReport) {
            return testStatsReport(pc2, statsReport);
          })]).then(t.step_func_done()).catch(t.step_func(function (err) {
            assert_unreached('test failed with error: ' + err);
          }));
        }, 200);
      });
  
      var onTrackCount = 0;
      var onDataChannelCalled = false;
  
      pc2.addEventListener('track', t.step_func(function () {
        onTrackCount++;
        if (onTrackCount === 2 && onDataChannelCalled) {
          onConnected();
        }
      }));
  
      pc2.addEventListener('datachannel', t.step_func(function () {
        onDataChannelCalled = true;
        if (onTrackCount === 2) {
          onConnected();
        }
      }));
  
      exchangeIceCandidates(pc1, pc2);
      doSignalingHandshake(pc1, pc2);
    })).catch(t.step_func(function (err) {
      assert_unreached('test failed with error: ' + err);
    }));
  }, 'getStats() with connected peer connections having tracks and data channel should return all mandatory to implement stats');

</script>
